Convert from nose to unittest.
So no one is going to take over nose and it is going to stop working in
either Python 3.9 or 3.10.  It turns out that it doesn't add much over
running plain unittest so I'm going to make a few changes to make that
possible.  The hardest part is learning to stop typing `nosetests`...