Replace type hint files with inline typing
This was required to make type checking work with the internal functions
that are scattered around the implementation.  I had to rewrite a few
things as well.

`rewrite_url` was given a bit of an overhaul.  I replaced `**kwargs`
with explicit parameters.  Since detecting whether the parameter is
specified or not is required, I went with a sentinel parameter and
chose the literal `'...'`.  It was the best way to make the type
signature readable.  If you use `'...'` in your URLs, I apologize.
You can set `ietfparse.algorithms.UNSPECIFIED` with something more
appropriate for your use case.

I did some minor rework in ordering helper function in `parse_accept`
as well.  It was returning a float instead of the int required by
`functools.cmp_to_key`.