Deprecate URL related functions
Use yarl instead.  It does a better job at manipulating URLs in every
way.  The only thing that you might want to do is define a helper for
sanitizing URLs if you find `url.with_password('***')` doesn't work
for you.  This will remove a lot of complexity from this library along
with the burden of maintaining it.