Replace setuptools with hatch
This seems to be the direction that the winds are blowing since hatch
was adopted by pypa in May 2022.  This is also the only option to
support editable installs and complete support for specifying metadata
in pyproject.yaml.

Version number management changed slightly since I'm going to try and
let hatch manage it.  It does this using a regex over a file and it
wants the version number in string form.  Instead of defining
``version_info`` and then combining it into ``version``, I'm defining
``version`` and splitting it into ``version_info``.