Fix the "all" hatch environment
Specifying an empty dependency list in the "all" environment wasn't the
right thing to do.  For some reason I thought that it was required and
an empty list was stating that no additional dependencies were required.
However, this created an environment without dependencies installed ...
this meant that the only command that worked in the "all" environment
was "test" since it has no additional dependencies.