Rework CI pipeline
I moved the guts of the CI pipeline into the "ci" hatch
environment. The only wrinkle is that the code-climate
test reporter is only used in the Circle CI pipeline;
however, the name of the coverage data file is on the
code-climate test reporter command line and controlled
by the hatch script. This is the only coupling between
the Circle CI config and pyproject.toml.