Remove support for developing with pip
I switched to hatch a while ago. This change removes the option to do a
`pip install -e '.[dev]'` altogether. Aside from personal preference,
using an "extra" for this is more of a hack than anything else since
it installs the package, its dependencies, and then the development
tools. This works for this project but isn't a good practice since it
locks development and installation dependencies together.